refactor(example): split demo screen into components and modules
Move the 1k-line app/index.tsx into a thin route, HomeScreen, shared styles/utils, and focused UI components so the example stays maintainable as a reference app.
