+++ example/src/components/Row.tsx
@@ -1,0 +1,12 @@
+import { Text, View } from "react-native";
+
+import { s } from "../styles";
+
+export function Row({ label, value }: { label: string; value: string }) {
+  return (
+    <View style={s.row}>
+      <Text style={s.rowLabel}>{label}</Text>
+      <Text style={s.rowValue}>{value}</Text>
+    </View>
+  );
+}
