+++ example/src/styles.ts
@@ -1,0 +1,97 @@
+import { StyleSheet } from "react-native";
+
+import { C } from "./theme";
+
+export const s = StyleSheet.create({
+  root: { flex: 1, backgroundColor: C.bg },
+  scroll: { padding: 24, paddingBottom: 48, flexGrow: 1 },
+  header: { alignItems: "center", marginBottom: 24, marginTop: 8 },
+  title: { color: C.white, fontSize: 26, fontWeight: "700", letterSpacing: -0.5 },
+  subtitle: { color: C.muted, fontSize: 13, marginTop: 4 },
+  badgeRow: { flexDirection: "row", alignItems: "center", justifyContent: "center", marginBottom: 8 },
+  dot: { width: 8, height: 8, borderRadius: 4, marginRight: 8 },
+  badgeText: { color: C.muted, fontSize: 13 },
+  connectionText: { color: C.white, fontSize: 13, textAlign: "center", marginBottom: 12 },
+  errorBox: { backgroundColor: C.errorBg, borderRadius: 10, padding: 14, marginBottom: 16 },
+  errorText: { color: C.error, fontSize: 13, lineHeight: 18 },
+  connectBtn: {
+    backgroundColor: C.green,
+    borderRadius: 32,
+    height: 52,
+    alignItems: "center",
+    justifyContent: "center",
+    marginTop: 16,
+  },
+  connectBtnText: { color: C.bg, fontWeight: "700", fontSize: 16 },
+  card: {
+    backgroundColor: C.surface,
+    borderRadius: 16,
+    padding: 20,
+    marginBottom: 12,
+    alignItems: "center",
+  },
+  avatar: { width: 80, height: 80, borderRadius: 40, marginBottom: 12 },
+  avatarFallback: { backgroundColor: "#333", alignItems: "center", justifyContent: "center" },
+  avatarInitial: { color: C.muted, fontSize: 28 },
+  profileName: { color: C.white, fontSize: 20, fontWeight: "700", marginBottom: 4 },
+  profileMeta: { color: C.muted, fontSize: 13, marginBottom: 2, textAlign: "center" },
+  cardTitle: {
+    color: C.white,
+    fontWeight: "700",
+    fontSize: 13,
+    alignSelf: "flex-start",
+    marginBottom: 8,
+    textTransform: "uppercase",
+    letterSpacing: 0.8,
+  },
+  row: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: "100%",
+    paddingVertical: 8,
+    borderTopWidth: StyleSheet.hairlineWidth,
+    borderTopColor: C.border,
+  },
+  rowLabel: { color: C.muted, fontSize: 13, flex: 1, marginRight: 8 },
+  rowValue: { color: C.white, fontSize: 13, fontWeight: "500" },
+  actions: { gap: 10, marginTop: 4 },
+  restoreHint: { color: C.muted, fontSize: 12, textAlign: "center", marginBottom: 10 },
+  emptyHint: { color: C.muted, fontSize: 13, textAlign: "center", marginTop: 8 },
+  nowPlayingRow: { flexDirection: "row", width: "100%", alignItems: "center", gap: 12 },
+  nowPlayingArtwork: { width: 64, height: 64, borderRadius: 10, backgroundColor: "#2b2b2b" },
+  nowPlayingArtworkPlaceholder: { alignItems: "center", justifyContent: "center" },
+  nowPlayingArtworkIcon: { color: C.muted, fontSize: 24 },
+  nowPlayingMeta: { flex: 1, minWidth: 0, alignItems: "flex-start" },
+  nowPlayingUri: { color: C.muted, fontSize: 11, marginTop: 2 },
+  transportRow: { flexDirection: "row", justifyContent: "space-between", marginTop: 10, width: "100%" },
+  librarySaveRow: { width: "100%", marginTop: 10 },
+  emptyBrowse: { alignItems: "center", justifyContent: "center", marginTop: 14, marginBottom: 8 },
+  emptyBrowseIcon: { color: C.green, fontSize: 28, marginBottom: 4 },
+  browseImage: { width: 90, height: 90, borderRadius: 8, marginTop: 12, marginBottom: 8 },
+  browseRow: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    width: "100%",
+    paddingVertical: 10,
+    borderTopWidth: StyleSheet.hairlineWidth,
+    borderTopColor: C.border,
+  },
+  btn: {
+    borderRadius: 32,
+    height: 48,
+    alignItems: "center",
+    justifyContent: "center",
+    borderWidth: 1.5,
+  },
+  btnFull: { width: "100%" },
+  transportBtn: {
+    width: "31%",
+    borderRadius: 32,
+    height: 48,
+    alignItems: "center",
+    justifyContent: "center",
+    borderWidth: 1.5,
+  },
+  btnText: { fontWeight: "600", fontSize: 15 },
+  disabledOpacity: { opacity: 0.5 },
+});
