+++ example/src/theme.ts
@@ -1,0 +1,10 @@
+export const C = {
+  bg: "#121212",
+  surface: "#1e1e1e",
+  border: "#2a2a2a",
+  green: "#1DB954",
+  white: "#ffffff",
+  muted: "#b3b3b3",
+  error: "#e25d5d",
+  errorBg: "#2d1515",
+} as const;
